pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "numpy>=2.3.2",
    "openai>=1.102.0",
    "pypdf2>=3.0.1",
    "streamlit>=1.49.0",
]
